move the code folder
